% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\name{CNA}
\alias{CNA}
\title{Create `Copy Number Array' data object}
\usage{
CNA(genomdat, chrom, maploc, data.type = c("logratio", "binary"),
  sampleid = NULL, presorted = FALSE)
}
\arguments{
\item{genomdat}{a vector or matrix of data from array-CGH, ROMA, or
other copy number experiments. If it is a matrix the rows correspond
to the markers and the columns to the samples.}

\item{chrom}{the chromosomes (or other group identifier) from which
the markers came.  Vector of length same as the number of rows of
genomdat.  If one wants the chromosomes to be ordered in the
natural order, this variable should be numeric or ordered category.}

\item{maploc}{the locations of marker on the genome.  Vector of length
same as the number of rows of genomdat. This has to be numeric.}

\item{data.type}{logratio (aCGH, ROMA, etc.) or binary (LOH).}

\item{sampleid}{sample identifier.  If missing the samples are named
by prefixing "Sample" to consecutive integers.}

\item{presorted}{logical indicator telling if the data have already
been sorted by chrom and maploc.  Default is FALSE.}
}
\value{
An object of class \code{CNA}
}
\description{
Creates a `copy number array' data object used for DNA copy number
analyses.
}
